Select 16-bit carriers on compression-enabled pins
Simpatico encodes 16-bit element widths, so the pin driver selects INT16 and
UINT16 carriers whatever the pin's compression settings are. This removes
floor_carrier_for_compression, which widened those two selections to 32 bits, and
the compression_will_run predicate that existed only to drive it -- both pin
drivers derived it, and it threaded through narrow_pin_chunk and
materialize_pin_batches. The three public entry points' signatures and
pin_materialization_options are unchanged.

A table whose values fit 16 bits now stores half the bytes it did per narrowed
column on a compression-enabled pin, on both tiers.

Tests. The INT64 operator sweep asserted only SUM(k), so it would have passed
even if every chunk silently failed to compress; the new narrowed-carrier sweep
censuses the pinned entry per operator instead, asserting both the recorded
carrier and compressed_chunks == chunks across all ten operators an integer plan
block can name, at each of the INT8, INT16 and INT32 carriers narrowing selects.
Its fixture repeats each value ten times because an all-distinct column expands
under rle and would fail the census for a reason unrelated to the carrier. A
second new case pins a native SMALLINT column with narrowing left off, which
reaches the compressor at INT16 through the chooser returning nothing rather than
through narrowing, and guards the 16-bit support on its own.

The host- and GPU-tier stacking tests and the MVCC delta test now expect INT16
where they expected INT32; the stacking test gained a UBIGINT column so UINT16 is
covered end to end. The width-explicit-op test keeps its subject -- a bitextract
field spec of a fixed total width still cannot encode a column narrowed to a
different width -- at the narrower carrier. The unit test for the deleted
function is removed; its chooser assertions duplicate the exhaustive pure-chooser
cases in the same file. One IS NOT DISTINCT FROM query joins the partition
suite's empty-side set, so the null-safe comparison runs over narrow payload
carriers.

Flag-off is unaffected: narrow_pin_chunk is called only under
enable_compressed_materialization, so with the flag off no code path here runs.

Verified: standalone Simpatico operator_sweep 16502/16502 chains (i16 1110/1110,
u16 1110/1110) and fused_operator_sweep 644/644 (i16 161/161).

+++ test/cpp/integration/test_pin_table_mvcc_insert.cpp
@@ -482,7 +482,7 @@ TEST_CASE_METHOD(PinMvccInsertFixture,
   namespace fs = std::filesystem;
 
   // Narrow-eligible values: they fit INT16, so the compression-enabled pin
-  // stores floored INT32 carriers inside compressed chunks.
+  // stores INT16 carriers inside compressed chunks.
   run_ok("CREATE TABLE t_delta_comp AS SELECT (range % 1000)::BIGINT AS k FROM range(50000);");
   run_ok("CHECKPOINT;");
 
@@ -514,7 +514,7 @@ TEST_CASE_METHOD(PinMvccInsertFixture,
   REQUIRE(census.compressed_chunks == census.chunks);
   REQUIRE(census.all_columns_narrowed);
   REQUIRE(census.first_chunk_carriers ==
-          std::vector<cudf::data_type>{cudf::data_type{cudf::type_id::INT32}});
+          std::vector<cudf::data_type>{cudf::data_type{cudf::type_id::INT16}});
 
   // Rows beyond the pinned prefix — one of them representable by no narrow
   // carrier. Delta splits decode fresh at native width, so the gate must
